learning rate 조절 정책을 반영 #58

+++ bin/train.py
@@ -14,7 +14,7 @@ __copyright__ = 'No copyright. Just copyleft!'
 ###########
 import argparse
 import codecs
-from collections import Counter
+import copy
 import logging
 import os
 import shutil
@@ -41,6 +41,44 @@ EMBED_DIM = 50
 GPU_NUM = 0
 BATCH_SIZE = 100
 EPOCH_NUM = 100
+RVT_ITER = 24    # 2 epoch
+RVT_TERM = 10
+
+
+#########
+# types #
+#########
+class CheckPoint(object):
+    """
+    check point to revert model and optimizer
+    """
+    def __init__(self, optimizer, model, state_dict):
+        self.state_dict = state_dict
+        self.model_dump = copy.deepcopy(state_dict)
+        self.model_dump['model'] = model.state_dict()
+        self.model_dump['optim'] = optimizer.state_dict()
+        self.optimizer = optimizer
+
+
+    def __str__(self):
+        lrs = [str(param_group['lr']) for param_group in self.optimizer.param_groups]
+        return '{state_dict: %s, lrs: %s}' % (self.state_dict, ', '.join(lrs))
+
+    def save(self, path):
+        """
+        save current state
+        :param  path:  path
+        """
+        torch.save(self.model_dump, path)
+
+    @classmethod
+    def load(cls, path):
+        """
+        load previous state
+        :param  path:  path
+        :return:  dumped model
+        """
+        return torch.load(path)
 
 
 #############
@@ -58,16 +96,21 @@ def _make_model_id(args):
     model_ids.append('e%d' % args.embed_dim)
     model_ids.append('gzte' if args.gazet_embed else 'gzt1')
     model_ids.append('pe%d' % (1 if args.pos_enc else 0))
+    model_ids.append('ri%d' % args.rvt_iter)
+    model_ids.append('rt%d' % args.rvt_term)
     return '.'.join(model_ids)
 
-
-def run(args):    # pylint: disable=too-many-locals,too-many-statements
+def _init(args):
     """
-    run function which is the start point of program
+    initialize dictionaries and model
     :param  args:  arguments
+    :return:  (vocabulary, gazetteer, data_, model) tuple
     """
     voca = data.load_voca(args.rsc_dir, args.phoneme, args.cutoff)
     gazet = gazetteer.load(codecs.open("%s/gazetteer.dic" % args.rsc_dir, 'r', encoding='UTF-8'))
+
+    # Load Data
+    data_ = data.load_data(args.in_pfx, voca)
 
     # Build Model
     if args.model_name.lower() == 'fnn5':
@@ -82,9 +125,17 @@ def run(args):    # pylint: disable=too-many-locals,too-many-statements
         model = models.Cnn7(args.window, voca, gazet,
                             args.embed_dim, hidden_dim,
                             args.phoneme, args.gazet_embed, args.pos_enc)
+    else:
+        raise ValueError('unknown model name: %s' % args.model_name)
+    return voca, gazet, data_, model
 
-    # Load Data
-    data_ = data.load_data(args.in_pfx, voca)
+
+def run(args):    # pylint: disable=too-many-locals,too-many-statements
+    """
+    run function which is the start point of program
+    :param  args:  arguments
+    """
+    voca, gazet, data_, model = _init(args)
 
     # Load GPU
     os.environ['CUDA_VISIBLE_DEVICES'] = str(args.gpu_num)
@@ -95,23 +146,42 @@ def run(args):    # pylint: disable=too-many-locals,too-many-statements
     criterion = nn.CrossEntropyLoss()
     optimizer = torch.optim.Adam(model.parameters())
 
-    model_id = _make_model_id(args)
-    logf = None
-    if args.logdir:
-        model_dir = '%s/%s' % (args.logdir, model_id)
-        if os.path.exists(model_dir):
-            logging.info('removing log: %s' % model_dir)
-            shutil.rmtree(model_dir)
-            time.sleep(3)
-        sum_wrt = SummaryWriter(model_dir)
-        logf = open('%s/%s.tsv' % (args.logdir, model_id), 'wt')
-        print('iter\tloss\taccuracy\tf-score', file=logf)
     losses = []
     accuracies = []
     f_scores = []
 
-    iter_ = 0
-    for epoch in range(args.epoch_num):
+    has_check_point = os.path.exists(args.output) and os.path.exists('%s.chk' % args.output)
+    if has_check_point:
+        #model.load(args.output)
+        model_dump = CheckPoint.load('%s.chk' % args.output)
+        model.load_state_dict(model_dump['model'])
+        optimizer.load_state_dict(model_dump['optim'])
+        best_iter = model_dump['iter']
+        losses.append(model_dump['loss'])
+        accuracies.append(model_dump['accuracy'])
+        f_scores.append(model_dump['f-score'])
+        iter_ = best_iter + 1
+        logging.info('==== reverting from check point ====')
+    else:
+        iter_ = 1
+        best_iter = 1
+
+    model_id = _make_model_id(args)
+    logf = None
+    if args.logdir:
+        model_dir = '%s/%s' % (args.logdir, model_id)
+        if os.path.exists(model_dir) and not has_check_point:
+            logging.info('==== removing log: %s ====', model_dir)
+            shutil.rmtree(model_dir)
+            time.sleep(3)
+        sum_wrt = SummaryWriter(model_dir)
+        log_path = '%s/%s.tsv' % (args.logdir, model_id)
+        logf = open(log_path, 'at' if has_check_point else 'wt')
+        if os.path.getsize(log_path) == 0:
+            print('iter\tloss\taccuracy\tf-score', file=logf)
+
+    revert = 0
+    while revert < args.rvt_term:
         for train_sent in data_['train']:
             # Convert to CUDA Variable
             train_labels, train_contexts, train_gazet = \
@@ -133,7 +203,6 @@ def run(args):    # pylint: disable=too-many-locals,too-many-statements
             loss = criterion(outputs, train_labels)
             loss.backward()
             optimizer.step()
-            iter_ += 1
 
             # Validation
             if iter_ % 1000 == 0:
@@ -145,12 +214,10 @@ def run(args):    # pylint: disable=too-many-locals,too-many-statements
                 losses.append(loss.data[0])
                 for dev_sent in data_['dev']:
                     # Convert to CUDA Variable
-                    _, dev_contexts, dev_gazet =\
-                            dev_sent.to_tensor(voca, gazet,
-                                               args.window, args.phoneme, args.gazet_embed)
+                    _, dev_contexts, dev_gazet = \
+                        dev_sent.to_tensor(voca, gazet, args.window, args.phoneme, args.gazet_embed)
                     dev_contexts = Variable(dev_contexts, volatile=True)
                     dev_gazet = Variable(dev_gazet, volatile=True)
-
                     if torch.cuda.is_available():
                         dev_contexts = dev_contexts.cuda()
                         dev_gazet = dev_gazet.cuda()
@@ -159,28 +226,51 @@ def run(args):    # pylint: disable=too-many-locals,too-many-statements
                     _, predicts = outputs.max(1)
                     dev_sent.compare_label(predicts, voca, measure)
 
-                accuracy_char, f_score = measure.get_score()
+                accuracy, f_score = measure.get_score()
                 print(file=sys.stderr)
                 sys.stderr.flush()
                 if not f_scores or f_score > max(f_scores):
-                    logging.info('writing best model..')
+                    logging.info('==== writing best model: %f ====', f_score)
                     model.save(args.output)
-                accuracies.append(accuracy_char)
+                    check_point = CheckPoint(optimizer, model,
+                                             {'iter': iter_, 'loss': loss.data[0],
+                                              'accuracy': accuracy, 'f-score': f_score})
+                    check_point.save('%s.chk' % args.output)
+                    logging.info('check point: %s', check_point)
+                    revert = 0
+                    best_iter = iter_
+                accuracies.append(accuracy)
                 f_scores.append(f_score)
-                logging.info('epoch: %d, iter: %dk, loss: %f, accuracy: %f, f-score: %f (max: %r)',
-                             epoch, iter_ // 1000, losses[-1], accuracy_char, f_score,
-                             max(f_scores))
+                logging.info('---- iter: %dk, loss: %f, accuracy: %f, f-score: %f (max: %r) ----',
+                             iter_ // 1000, losses[-1], accuracy, f_score, max(f_scores))
                 if args.logdir:
                     sum_wrt.add_scalar('loss', losses[-1], iter_ // 1000)
-                    sum_wrt.add_scalar('accuracy', accuracy_char, iter_ // 1000)
+                    sum_wrt.add_scalar('accuracy', accuracy, iter_ // 1000)
                     sum_wrt.add_scalar('f-score', f_score, iter_ // 1000)
-                    print('{}\t{}\t{}\t{}'.format(iter_ // 1000, losses[-1], accuracy_char,
+                    print('{}\t{}\t{}\t{}'.format(iter_ // 1000, losses[-1], accuracy,
                                                   f_score), file=logf)
                     logf.flush()
+
+                # revert policy
+                if (iter_ - best_iter) > (args.rvt_iter * 1000):
+                    #model.load(args.output)
+                    model_dump = CheckPoint.load('%s.chk' % args.output)
+                    model.load_state_dict(model_dump['model'])
+                    optimizer.load_state_dict(model_dump['optim'])
+                    lrs = []
+                    for param_group in optimizer.param_groups:
+                        param_group['lr'] *= 0.9
+                        lrs.append(param_group['lr'])
+                    best_iter = iter_
+                    revert += 1
+                    logging.info('==== revert to iter: %dk, revert count: %d ====', iter_ // 1000,
+                                 revert)
+                    logging.info('learning rates: %s', ', '.join([str(_) for _ in lrs]))
             elif iter_ % 100 == 0:
                 print('.', end='', file=sys.stderr)
                 sys.stderr.flush()
 
+            iter_ += 1
 
 ########
 # main #
@@ -204,15 +294,17 @@ def main():
                         metavar='INT', type=int, default=GPU_NUM)
     parser.add_argument('--batch-size', help='batch size <default: %d>' % BATCH_SIZE, metavar='INT',
                         type=int, default=BATCH_SIZE)
-    parser.add_argument('--epoch-num', help='epoch number <default: %d>' % EPOCH_NUM, metavar='INT',
-                        type=int, default=EPOCH_NUM)
     parser.add_argument('--phoneme', help='expand phonemes context', action='store_true')
     parser.add_argument('--pos-enc', help='add positional encoding',
                         action='store_true', default=False)
     parser.add_argument('--gazet-embed', help='gazetteer type', action='store_true',
                         default=False)
     parser.add_argument('--cutoff', help='cutoff', action='store',\
-            type=int, metavar="int", default=5)
+                        type=int, metavar='NUM', default=5)
+    parser.add_argument('--rvt-iter', help='최대치 파라미터로 되돌아갈 반복 횟수 (천단위) <default: %d>' % \
+                                           RVT_ITER, type=int, metavar='NUM', default=RVT_ITER)
+    parser.add_argument('--rvt-term', help='파라미터로 되돌아갈 최대 횟수 <default: %d>' % \
+                                           RVT_TERM, type=int, metavar='NUM', default=RVT_TERM)
     parser.add_argument('--debug', help='enable debug', action='store_true')
     args = parser.parse_args()
